+++ gebeh-web/Cargo.toml
@@ -16,6 +16,7 @@ crate-type = ["cdylib"]
 wasm-bindgen = "0.2.45"
 gebeh-core = { path = "../gebeh-core" }
 gebeh-front-helper = { path = "../gebeh-front-helper" }
+gebeh-network = { path = "../gebeh-network" }
 web-time = "1.1.0"
 wee_alloc = "0.4.5"
 rkyv = "0.8.15"
